requirements.txt
autobahn
cbor
psutil
speedtest-cli
txaio
